+++ scratch.js
@@ -1,26 +1,42 @@
-'use strict';
 const wtf = require('./src/index');
-// const fromFile = require('./tests/lib/_cachedPage');
+const readFile = require('./tests/lib/_cachedPage');
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
 
-// todo:
-// description(), extract(), summary()
+// var str = `{{tag|ref|content=haha}}`;
+// var str = `{{Sfn|Tattersall|1982|pp=43–46}}`;
+// var str = `{{MSW3 | id = 13801049 | pages = 391–393 | heading = Genus ''Nycteris'' | author = Simmons, N. B.}}`;
+// var str = `{{Buddhist crisis|state=collapsed}}`;
+// var str = `{{Lacking ISBN|date=January 2017}}`;
+// var str = `The ring-tailed lemur is known locally in Malagasy as ''{{lang|mg|maky}}'' (pronounced {{IPA-mg|ˈmakʲi̥|}}), `;
+// var str = `{{s-ttl | title = Member of the [[List of United States Representatives from Massachusetts|House of Representatives]] <br /> from [[Massachusetts's 11th congressional district]] | years = 1947–1953 }}`;
+// var str = `{{CongBio|K000107}}`;
+// var str = `{{Gutenberg author | id=James,+Henry+(1843-1916) |name=Henry James}}`;
+// var str = `{{Internet Archive author |sname=Gyula Krúdy|birth=1878|death=1933}}`;
+// var str = `{{Goodreads author|1599723.Michael_Grant|Michael Grant}}`;
+// var str = `{{IMDb title | 0426883 | Alpha Dog }}`;
+// var str = `{{Goodreads book|140397|Wings of the Falcon}}`;
+// var str = `{{Goodreads book|id=140397|title=Wings of the Falcon}}`;
+// var str = `{{Internet Archive author |sname=John Fitzgerald [[Kennedy]] |sopt=t}}`;
+// var str = `{{Librivox asdf |id=2572}}`;
+// var str = `{{Librivox book |stitle=The Federalist Papers |dtitle=''The Federalist'' papers}}`;
+// var str = `{{discogs artist|artist=John F. Kennedy|date=june 9}}`;
+// str = `{{Discogs artist|artist=소녀시대|name=소녀시대}}`;
+// str = `{{Discogs artist|소녀시대|소녀시대}}`;
+// var str = `{{Find a Grave|574|accessdate=November 17, 2013}}`;
+// var str = `{{Dmoz|Society/History/By_Region/North_America/United_States/Presidents/Kennedy%2C_John_Fitzgerald/}}`;
+// var str = `{{iMDb name|0448123}}`;
+// var str = `{{Twitter | AcadiaU | Acadia University }}`;
+// var str = `{{Facebook|zuck|Mark Zuckerberg}}`;
+// var str = `{{ESPN NFL | 10536 | Trent Edwards }}`;
+// var str = `{{small|(1976–77)}}`;
+// var doc = wtf(str);
+// console.log(doc.plaintext());
+// console.log(doc.templates(0));
 
-//doc.infoboxes('Venue')
+// var doc = readFile('toronto');
 
-// let doc = fromFile('royal_cinema');
-// let doc = fromFile('toronto');
-// console.log(doc.sentences(0).text());
-// console.log(doc.sections('Infrastructure').json());
-// // wtf.fetch('Royal Cinema').then((doc) => {
-// //   console.log(doc.json({
-// //     categories: false
-// //   }));
-// // }).catch(console.log);
-
-// console.log(wtf(wiki).json({
-//   images: true
-// }));
-let wiki = `'''Toronto''' ({{IPAc-en|t|ɵ|ˈ|r|ɒ|n|t|oʊ}}, {{IPAc-en|local|ˈ|t|r|ɒ|n|oʊ}}) is the [[List of the 100 largest municipalities in Canada by population|most populous city]] in [[Canada]] and the [[Provinces and territories of Canada|provincial]] [[capital city|capital]] of [[Ontario]]. It is located in [[Southern Ontario]] on the northwestern shore of [[Lake Ontario]]. The [[history of Toronto]] began in the late 18th century when the [[The Crown|British Crown]] [[Toronto Purchase|purchased]] its land from the [[Mississaugas of the New Credit`;
-console.log(wtf(wiki).plaintext());
+wtf.fetch('Peter Mansbridge').then(doc => {
+  console.log(doc.plaintext());
+// console.log(doc.templates().filter(t => t.template !== 'citation'));
+});