+++ scripts/build.js
@@ -21,7 +21,8 @@ echo(banner).to(uncompressed);
 echo(banner).to(compressed);
 
 //browserify + derequire
-var cmd = browserify + ' ./src/index.js --standalone wtf';
+//var cmd = browserify + ' ./src/index.js --standalone wtf';
+var cmd = browserify + ' ./src/main.js --standalone wtf';
 cmd += ' -t [ babelify --presets [ env ] ]';
 cmd += ' | ' + derequire;
 cmd += ' >> ' + uncompressed;