+++ src/templates/dates/parsers.js
@@ -2,8 +2,22 @@ const dates = require('./dates');
 const ymd = dates.ymd;
 const toText = dates.toText;
 const delta = require('./delta_date');
+const strip = function(tmpl) {
+  tmpl = tmpl.replace(/^\{\{/, '');
+  tmpl = tmpl.replace(/\}\}$/, '');
+  return tmpl;
+};
+
+//wrap it up as a template
+const template = function(date) {
+  return {
+    template: 'date',
+    data: date
+  };
+};
 
 const getBoth = function(tmpl) {
+  tmpl = strip(tmpl);
   let arr = tmpl.split('|');
   let from = ymd(arr.slice(1, 4));
   let to = arr.slice(4, 7);
@@ -22,7 +36,8 @@ const getBoth = function(tmpl) {
 const parsers = {
 
   //generic {{date|year|month|date}} template
-  date: (tmpl, obj) => {
+  date: (tmpl, r) => {
+    tmpl = strip(tmpl);
     let arr = tmpl.split('|');
     arr = arr.slice(1, 8);
     //support 'df=yes|1894|7|26'
@@ -32,14 +47,14 @@ const parsers = {
     let date = ymd(arr);
     date.text = toText(date); //make the replacement string
     if (date.text) {
-      obj.dates = obj.dates || [];
-      obj.dates.push(date);
+      r.templates.push(template(date));
     }
     return date.text;
   },
 
   //support parsing of 'February 10, 1992'
-  natural_date: (tmpl, obj) => {
+  natural_date: (tmpl, r) => {
+    tmpl = strip(tmpl);
     let arr = tmpl.split('|');
     let str = arr[1] || '';
     // - just a year
@@ -57,36 +72,34 @@ const parsers = {
         date.date = d.getDate();
       }
     }
-    obj.dates = obj.dates || [];
-    obj.dates.push(date);
+    r.templates.push(template(date));
     return str.trim();
   },
 
   //just grab the first value, and assume it's a year
-  one_year: (tmpl, obj) => {
+  one_year: (tmpl, r) => {
+    tmpl = strip(tmpl);
     let arr = tmpl.split('|');
     let str = arr[1] || '';
     let year = parseInt(str, 10);
-    obj.dates = obj.dates || [];
-    obj.dates.push({
+    r.templates.push(template({
       year: year
-    });
+    }));
     return str.trim();
   },
 
   //assume 'y|m|d' | 'y|m|d'
-  two_dates: (tmpl, obj) => {
+  two_dates: (tmpl, r) => {
+    tmpl = strip(tmpl);
     let arr = tmpl.split('|');
     //'b' means show birth-date, otherwise show death-date
     if (arr[1] === 'B' || arr[1] === 'b') {
       let date = ymd(arr.slice(2, 5));
-      obj.dates = obj.dates || [];
-      obj.dates.push(date);
+      r.templates.push(template(date));
       return toText(date);
     }
     let date = ymd(arr.slice(5, 8));
-    obj.dates = obj.dates || [];
-    obj.dates.push(date);
+    r.templates.push(template(date));
     return toText(date);
   },