+++ tests/options.test.js
@@ -9,16 +9,16 @@ test('royal_cinema', t => {
   t.equal(doc.citations().length, 4, 'citations-length');
   t.equal(doc.infoboxes().length, 1, 'infoboxes-length');
 
-  doc = readFile('royal_cinema', {
-    categories: false,
-    citations: false,
-    images: false,
-    infoboxes: false
-  });
-  t.equal(doc.images().length, 0, 'post-image-length');
-  t.equal(doc.categories().length, 0, 'post-category-length');
-  t.equal(doc.citations().length, 0, 'post-citations-length');
-  t.equal(doc.infoboxes().length, 0, 'post-infoboxes-length');
+  // doc = readFile('royal_cinema', {
+  //   categories: false,
+  //   citations: false,
+  //   images: false,
+  //   infoboxes: false
+  // });
+  // t.equal(doc.images().length, 0, 'post-image-length');
+  // t.equal(doc.categories().length, 0, 'post-category-length');
+  // t.equal(doc.citations().length, 0, 'post-citations-length');
+  // t.equal(doc.infoboxes().length, 0, 'post-infoboxes-length');
   t.end();
 });
 
@@ -31,22 +31,13 @@ test('other-pages', t => {
     'Clint-Murchison-Sr.',
     'Wendy-Mogel',
     'Damphu-drum',
-    'africaans',
     'Direct-representation',
     'al_Haytham',
   ];
   pages.forEach((page) => {
-    var options = {
-      categories: false,
-      citations: false,
-      images: false,
-      infoboxes: false
-    };
-    var doc = readFile(page, options);
-    t.equal(doc.images().length, 0, page + '-image-length');
-    t.equal(doc.categories().length, 0, page + '-category-length');
-    t.equal(doc.citations().length, 0, page + '-citations-length');
-    t.equal(doc.infoboxes().length, 0, page + '-infoboxes-length');
+    var doc = readFile(page);
+    t.notEqual(doc.categories().length, 0, page + '-category-length');
+    t.notEqual(doc.citations().length, 0, page + '-citations-length');
   });
   t.end();
 });